third_party/blink/web_tests/fast/forms/label/label-event-order-expected.txt

Event order for a labeled control should be "focus" then "change"

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

"focus" event has been dispatched.
PASS changeCount is 0
"change" event has been dispatched.
PASS focusCount is 1
PASS successfullyParsed is true

TEST COMPLETE
